more changes

+++ app/Http/Controllers/CategoriesController.php
@@ -96,6 +96,11 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if($category->posts->count() > 0)
+        {
+            session()->flash("error", "category cannot be deleted because it has a post.");
+            return redirect()->back();
+        }
         $category->delete();
         session()->flash("success", "Category Deleted Successfully");
