+++ app/Http/Controllers/TagsController.php
@@ -96,6 +96,11 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if($tag->posts->count() > 0)
+        {
+            session()->flash("error","Tag cannot be deleted because it has posts");
+            return redirect()->back();
+        }
         $tag->delete();
         session()->flash("success", "Tag Deleted Successfully");
